Fix M2Track size, remove old ADT value

+++ WoWFormatLib/Utils/ABlock.cs
@@ -17,7 +17,7 @@ namespace WoWFormatLib
     public struct ArrayReference<T> where T : struct
     {
         public uint Number;
-        private long elementsOffset;
+        private uint elementsOffset;
 
         public IEnumerable<T> GetElements(BinaryReader bin)
         {
